Crate audio_thread_priority

audio_thread_priority

Promote the current thread, or another thread (possibly in another process), to real-time priority, suitable for low-latency audio processing.

Example

use audio_thread_priority::{promote_current_thread_to_real_time, demote_current_thread_from_real_time};

// ... on a thread that will compute audio and has to be real-time:
match promote_current_thread_to_real_time(512, 44100) {
  Ok(h) => {
    println!("this thread is now bumped to real-time priority.");

    // Do some real-time work...

    match demote_current_thread_from_real_time(h) {
      Ok(_) => {
        println!("this thread is now bumped back to normal.")
      }
      Err(_) => {
        println!("Could not bring the thread back to normal priority.")
      }
    };
  }
  Err(e) => {
    eprintln!("Error promoting thread to real-time: {}", e);
  }
}

Structs

AudioThreadPriorityError

The OS-specific issue is available as inner

atp_handle

Opaque handle for the C API

Functions

atp_demote_current_thread_from_real_time⚠

Demotes the calling thread from real-time priority, with a C API.

atp_free_handle⚠

Frees a handle, with a C API.

atp_promote_current_thread_to_real_time

Promote the calling thread thread to real-time priority, with a C API.

demote_current_thread_from_real_time

Demotes the calling thread from real-time priority.

promote_current_thread_to_real_time

Promote the calling thread thread to real-time priority.

Type Aliases

RtPriorityHandle

Opaque handle to a thread handle structure.